Move request sending to client sub module
