+++ crates/lune-std-net/src/shared/request.rs
@@ -2,32 +2,22 @@ use http_body_util::Full;
 use url::Url;
 
 use hyper::{
-    body::{Bytes, Incoming},
-    client::conn::http1::handshake,
-    header::{HeaderName, HeaderValue, LOCATION, USER_AGENT},
-    HeaderMap, Method, Request as HyperRequest, Response as HyperResponse, Uri,
+    body::Bytes,
+    header::{HeaderName, HeaderValue, USER_AGENT},
+    HeaderMap, Request as HyperRequest,
 };
 
 use mlua::prelude::*;
 
-use crate::{
-    client::{config::RequestConfig, stream::HttpRequestStream},
-    shared::{
-        headers::create_user_agent_header,
-        hyper::{HyperExecutor, HyperIo},
-        response::Response,
-    },
-};
-
-const MAX_REDIRECTS: usize = 10;
+use crate::{client::config::RequestConfig, shared::headers::create_user_agent_header};
 
 #[derive(Debug, Clone)]
 pub struct Request {
     // NOTE: We use Bytes instead of Full<Bytes> to avoid
     // needing async when getting a reference to the body
-    inner: HyperRequest<Bytes>,
-    redirects: usize,
-    decompress: bool,
+    pub(crate) inner: HyperRequest<Bytes>,
+    pub(crate) redirects: usize,
+    pub(crate) decompress: bool,
 }
 
 impl Request {
@@ -100,43 +90,6 @@ impl Request {
         let body = Full::new(self.inner.body().clone());
         builder.body(body).expect("request was valid")
     }
-
-    /**
-        Sends the request and returns the final response.
-
-        This will follow any redirects returned by the server,
-        modifying the request method and body as necessary.
-    */
-    pub async fn send(mut self, lua: Lua) -> LuaResult<Response> {
-        loop {
-            let stream = HttpRequestStream::connect(self.inner.uri()).await?;
-
-            let (mut sender, conn) = handshake(HyperIo::from(stream)).await.into_lua_err()?;
-
-            HyperExecutor::execute(lua.clone(), conn);
-
-            let incoming = sender.send_request(self.as_full()).await.into_lua_err()?;
-
-            if let Some((new_method, new_uri)) = check_redirect(&self.inner, &incoming) {
-                if self.redirects >= MAX_REDIRECTS {
-                    return Err(LuaError::external("Too many redirects"));
-                }
-
-                if new_method == Method::GET {
-                    *self.inner.body_mut() = Bytes::new();
-                }
-
-                *self.inner.method_mut() = new_method;
-                *self.inner.uri_mut() = new_uri;
-
-                self.redirects += 1;
-
-                continue;
-            }
-
-            break Response::from_incoming(incoming, self.decompress).await;
-        }
-    }
 }
 
 fn add_default_headers(lua: &Lua, headers: &mut HeaderMap) -> LuaResult<()> {
@@ -147,24 +100,4 @@ fn add_default_headers(lua: &Lua, headers: &mut HeaderMap) -> LuaResult<()> {
     }
 
     Ok(())
-}
-
-fn check_redirect(
-    request: &HyperRequest<Bytes>,
-    response: &HyperResponse<Incoming>,
-) -> Option<(Method, Uri)> {
-    if !response.status().is_redirection() {
-        return None;
-    }
-
-    let location = response.headers().get(LOCATION)?;
-    let location = location.to_str().ok()?;
-    let location = location.parse().ok()?;
-
-    let method = match response.status().as_u16() {
-        301..=303 => Method::GET,
-        _ => request.method().clone(),
-    };
-
-    Some((method, location))
 }
